ai/requirements.txt
Flask==3.1.1
Flask-Cors==6.0.1
langchain-google-vertexai==2.0.27
langchain-community==0.3.27
langchain-huggingface==0.3.1
langchain-core==0.3.72
pypdf==5.9.0
chromadb==1.0.15
sentence-transformers==2.7.0
google-cloud-aiplatform==1.105.0
google-cloud-translate==3.13.0
google-auth==2.40.3
google-api-core==2.25.1
google-cloud-core==2.4.3
gunicorn==23.0.0
torch==2.2.2
numpy==1.26.4
-f https://download.pytorch.org/whl/cpu/torch_stable.html
